confl/python/requirements.txt
mysql-connector-python==8.0.28 
confluent_kafka
requests
